# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ProductMetafield": {
				"description": "Product metafield (custom field from Shopify)",
				"properties": {
					"namespace": {
						"type": "string"
					},
					"key": {
						"type": "string"
					},
					"value": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"namespace",
					"key",
					"value",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CartLineItem": {
				"description": "Cart line item details",
				"properties": {
					"id": {
						"type": "string"
					},
					"merchandiseId": {
						"type": "string"
					},
					"productId": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"variantTitle": {
						"type": "string"
					},
					"image": {
						"type": "string",
						"nullable": true
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"compareAtPrice": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"quantityAvailable": {
						"type": "number",
						"format": "double"
					},
					"metafields": {
						"items": {
							"$ref": "#/components/schemas/ProductMetafield"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"merchandiseId",
					"productId",
					"title",
					"variantTitle",
					"image",
					"price",
					"compareAtPrice",
					"quantity",
					"quantityAvailable",
					"metafields"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CartData": {
				"description": "Cart data",
				"properties": {
					"cartId": {
						"type": "string"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/CartLineItem"
						},
						"type": "array"
					},
					"totalQuantity": {
						"type": "number",
						"format": "double"
					},
					"subtotal": {
						"type": "number",
						"format": "double"
					},
					"currency": {
						"type": "string"
					}
				},
				"required": [
					"cartId",
					"items",
					"totalQuantity",
					"subtotal",
					"currency"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CartResponse": {
				"description": "Cart API response",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/CartData"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdateItemQuantityRequest": {
				"description": "Update item quantity request body",
				"properties": {
					"quantity": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Cartaisy API",
		"version": "1.0.0",
		"description": "E-commerce backend API with Shopify integration"
	},
	"paths": {
...
		"/cart/{cartId}/items/{lineItemId}": {
			"put": {
				"operationId": "UpdateItemQuantity",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CartResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"404": {
						"description": "Cart or item not found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Update cart item quantity",
				"tags": [
					"Cart"
				],
				"security": [],
				"parameters": [
					{
						"description": "- Shopify cart ID",
						"in": "path",
						"name": "cartId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "- Shopify line item ID",
						"in": "path",
						"name": "lineItemId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- New quantity",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateItemQuantityRequest",
								"description": "- New quantity"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RemoveItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CartResponse"
								}
							}
						}
					},
					"404": {
						"description": "Cart or item not found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Remove item from cart",
				"tags": [
					"Cart"
				],
				"security": [],
				"parameters": [
					{
						"description": "- Shopify cart ID",
						"in": "path",
						"name": "cartId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "- Shopify line item ID",
						"in": "path",
						"name": "lineItemId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/api/v1",
			"description": "Development server"
		}
	]
}